refactor: reorganize packages structure
