--- githubservice/pubspecyamlfetcher.go
@@ -1,9 +1,0 @@
-package githubservice
-
-type pubspecyamlFetcher struct {
-	*YamlFetcher
-}
-
-func (pubspecyamlFetcher *pubspecyamlFetcher) GetVersionUsingDefaultPath(ghContentProvider contentProvider) (string, error) {
-	return pubspecyamlFetcher.GetVersion(ghContentProvider, AtcSettings{Path: "pubspec.yaml"})
-}
